# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hermes Automation API
  version: 1.0.0
  summary: REST API for triggering and monitoring TISS billing automations.
  description: |
    External-system entry point to the Hermes automation pipeline. A caller can:

    - upload one or more TISS files (`.zip` / `.xml`) to create a job;
    - optionally pre-supply a `platformCredentialId` to skip the human approval step;
    - poll job status or stream live events via SSE;
    - manage platform credentials programmatically.

    All endpoints are scoped per API key owner. Resources are isolated per user; an
    API key cannot read or modify resources owned by another user.
  contact:
    name: Webstar
    email: dev@webstar.studio
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://hermes.example.com
    description: Production (replace with your deployment)
...
security:
  - apiKey: []
paths:
...
  /api/v1/platform-credentials:
    get:
      tags: [Platform credentials]
      operationId: listPlatformCredentials
      summary: List credentials owned by the API key holder
      responses:
        "200":
          description: Credential list (newest first).
          content:
            application/json:
              schema:
                type: object
                required: [ok, credentials]
                properties:
                  ok: { const: true }
                  credentials:
                    type: array
                    items: { $ref: "#/components/schemas/PlatformCredential" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Platform credentials]
      operationId: createPlatformCredential
      summary: Create a credential
      description: |
        The password is encrypted server-side using AES-256-GCM and is never returned
        in any response. The plain `username` is also masked when read back
        (`usernameMasked` is the only field exposed by API).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialCreate" }
      responses:
        "200":
          description: Credential created.
          content:
            application/json:
              schema:
                type: object
                required: [ok, credential]
                properties:
                  ok: { const: true }
                  credential: { $ref: "#/components/schemas/PlatformCredential" }
        "400":
          description: Invalid body (missing fields, value out of range).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: A credential with this label already exists for the platform.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: hapi_*
      description: |
        API key issued under **Settings → API Keys** (`/app/settings/api-keys`).

        Format: `hapi_` + 32-character base32 body (≈ 37 chars total). The plaintext
        secret is shown **only once** at creation; the server stores `sha256(secret)`
        and discards the raw value.

        Send the key as `Authorization: Bearer hapi_xxxxxxxxxx…`.
...
  responses:
    Unauthorized:
      description: Missing, malformed, revoked, or expired API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: API key owner is not in `approved` status.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    PlatformId:
      type: string
      enum: [orizon_fature]
      default: orizon_fature
...
    PlatformCredential:
      type: object
      required: [id, platformId, label, usernameMasked]
      properties:
        id: { type: string, format: uuid }
        platformId: { $ref: "#/components/schemas/PlatformId" }
        label: { type: string, minLength: 2, maxLength: 80 }
        usernameMasked:
          type: string
          description: Username with the middle masked (e.g. `us***om`). The plaintext is never returned.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CredentialCreate:
      type: object
      required: [platformId, label, username, password]
      properties:
        platformId: { $ref: "#/components/schemas/PlatformId" }
        label: { type: string, minLength: 2, maxLength: 80 }
        username: { type: string, minLength: 1, maxLength: 160 }
        password:
          type: string
          minLength: 1
          maxLength: 512
          format: password
...
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok: { const: false }
        error:
          type: object
          required: [message]
          properties:
            code:
              type: string
              description: |
                Stable, machine-readable code. Known values include:
                `UNAUTHORIZED`, `PENDING_APPROVAL`, `NOT_FOUND`, `INVALID_BODY`,
                `FILE_REQUIRED`, `TOO_MANY_FILES`, `UPLOAD_FAILED`,
                `CREDENTIAL_NOT_FOUND`, `DUPLICATE_LABEL`, `CREDENTIAL_IN_USE`,
                `RUN_NOT_READY`, `RUN_NOT_FOUND`.
            message: { type: string }